feat(dapi): moved the sendProviderConnected function to a more appropriate place
refs. metriport/metriport#665
